Bug #11397: check that numeric --orientation arguments are in range
The only valid parameters to -o (--orientation) are 0, 1, 2, 3, normal, left,
inverted, and right.  xrandr converts the strings to numbers and then checks
that they're within range, but doesn't validate them if it was numeric to begin
with.

Move the range check outside of the if statement so that out-of-range numeric
values are rejected properly.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: Daniel Dadap <ddadap@nvidia.com>